Fixes for things found in Travis
1. Fix 32bit compiling in pack.cc
2. Fix valgrind unitialized jump in statement.cc
3. Fix valgrind cleanup
4. Fix 2 scope issues with cppcheck
5. Fix docs for older versions of Sphinx

+++ src/ascore/pack.cc
@@ -105,7 +105,7 @@ char *ascore_pack_length(char *buffer, size_t length)
   {
     buffer[0]= 0xfe;
     buffer++;
-    ascore_pack_int8(buffer, length);
+    ascore_pack_int8(buffer, (uint64_t)length);
     buffer+= 8;
   }
